+++ vyper/evm/assembler/symbols.py
@@ -2,11 +2,13 @@ from typing import Any, TypeVar
 
 from vyper.evm.assembler.instructions import (
     CONST,
+    CONSTREF,
     DATA_ITEM,
     PUSH_OFST,
     PUSHLABEL,
     AssemblyInstruction,
     DataHeader,
+    Label,
     TaggedInstruction,
     calc_push_size,
 )
@@ -15,46 +17,6 @@ from vyper.exceptions import CompilerPanic
 from vyper.utils import OrderedSet
 
 SYMBOL_SIZE = 2  # size of a PUSH instruction for a code symbol
-
-
-class Label:
-    def __init__(self, label: str):
-        assert isinstance(label, str)
-        self.label = label
-
-    def __repr__(self):
-        return f"LABEL {self.label}"
-
-    def __eq__(self, other):
-        if not isinstance(other, Label):
-            return False
-        return self.label == other.label
-
-    def __hash__(self):
-        return hash(self.label)
-
-
-def is_label(i):
-    return isinstance(i, Label)
-
-
-# this could be fused with Label, the only difference is if
-# it gets looked up from const_map or symbol_map.
-class CONSTREF:
-    def __init__(self, label: str):
-        assert isinstance(label, str)
-        self.label = label
-
-    def __repr__(self):
-        return f"CONSTREF {self.label}"
-
-    def __eq__(self, other):
-        if not isinstance(other, CONSTREF):
-            return False
-        return self.label == other.label
-
-    def __hash__(self):
-        return hash(self.label)
 
 
 T = TypeVar("T")
